% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CURT_SurvivalRegressionTree.R
\name{curt}
\alias{curt}
\title{Survival analysis using censoring unbiased random forests}
\usage{
curt(
  formula,
  train_data,
  test_data,
  n.tree = 1000,
  tau1 = NULL,
  type = "drl",
  mtry = NULL
)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{train_data}{A dataframe containing the training dataset.}

\item{test_data}{A dataframe containing the test dataset for which survival predictions will be made for.}

\item{n.tree}{Number of trees used in the forest}

\item{tau1}{At what time-point survival predictions should be calculated (e.g, if time is measured in years and tau=1 the algorithm would predict one year survival probabilities).}

\item{type}{The type of loss function to be used in the random forest algorithm. Type = "bjl" uses the Buckley-James loss function and is the default and type= "drl" uses the doubly robust loss function.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.}
}
\value{
A list containing survival predictions on the test set and variable importance measures
}
\description{
This function calculates survival predictions (e.g., one year survival probability) and variable importance measures using censoring unbiased random forests for both Buckley-James and doubly robust loss functions.
}
\examples{
train_data_path <- system.file("extdata", "train_data.csv", package = "CURT")
test_data_path <- system.file("extdata", "test_data.csv", package = "CURT")
train_data <- read.csv(train_data_path)
test_data <- read.csv(test_data_path)
result <- curt(Surv(obs, delta) ~ . -x2-x3, train_data, test_data, n.tree=500, tau1=0.6, type="drl")
}
